pyproject.toml
[project]
name = "qqq-law-engine"
version = "0.1.0"
description = "QQQ Cycle-Conditional Law Engine per SRD v8.7 and ADD v1.0."
requires-python = ">=3.11,<3.12"
license = { text = "Proprietary" }
authors = [{ name = "Quantitative Finance System Maintainers" }]
keywords = ["quantitative-finance", "point-in-time", "backtest", "qqq"]
dependencies = [
    "cvxpy>=1.5,<1.7",
    "ecos>=2.0,<3.0",
    "httpx>=0.27,<0.29",
    "jsonschema>=4.23,<5.0",
    "numpy>=1.26,<3.0",
    "orjson>=3.10,<4.0",
    "pandas>=2.2,<3.0",
    "pyarrow>=15,<22",
    "pydantic>=2.8,<3.0",
    "pydantic-settings>=2.4,<3.0",
    "python-dateutil>=2.9,<3.0",
    "pyyaml>=6.0,<7.0",
    "scipy>=1.13,<2.0",
    "typer>=0.12,<1.0",
]

[project.optional-dependencies]
dev = [
    "coverage[toml]>=7.6,<8.0",
    "hypothesis>=6.112,<7.0",
    "import-linter>=2.0,<3.0",
    "jupyterlab>=4.2,<5.0",
    "mypy>=1.11,<2.0",
    "pandas-stubs>=2.2,<3.0",
    "pytest>=8.3,<9.0",
    "pytest-cov>=5.0,<7.0",
    "pytest-shard>=0.1,<1.0",
    "pytest-xdist>=3.6,<4.0",
    "ruff>=0.6,<0.7",
    "types-jsonschema>=4.23,<5.0",
    "types-python-dateutil>=2.9,<3.0",
    "types-pyyaml>=6.0,<7.0",
]

[build-system]
requires = ["hatchling>=1.25,<2.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/app", "src/backtest", "src/data_contract", "src/decision", "src/features", "src/inference", "src/law", "src/research", "src/state"]

[tool.uv]
package = false

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests", "tools"]
extend-exclude = [
    "artifacts",
    "data",
    "htmlcov",
    "notebooks",
]

[tool.ruff.lint]
select = [
    "A",
    "ARG",
    "B",
    "C4",
    "DTZ",
    "E",
    "F",
    "FA",
    "FBT",
    "I",
    "ISC",
    "N",
    "NPY",
    "PERF",
    "PIE",
    "PL",
    "PT",
    "PTH",
    "RET",
    "RUF",
    "S",
    "SIM",
    "T20",
    "UP",
    "W",
]
ignore = [
    "COM812",
    "ISC001",
]
unfixable = [
    "F401",
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101",
    "PLR2004",
]
"tools/**/*.py" = [
    "T201",
]

[tool.ruff.lint.isort]
known-first-party = ["app", "backtest", "data_contract", "decision", "features", "inference", "law", "research", "state"]
combine-as-imports = true
force-sort-within-sections = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
warn_unreachable = true
show_error_codes = true
pretty = true
mypy_path = "src"
explicit_package_bases = true
packages = [
    "app",
    "backtest",
    "data_contract",
    "decision",
    "features",
    "inference",
    "law",
    "research",
    "state",
]
files = ["tests", "tools"]
exclude = [
    "^data/",
    "^artifacts/",
    "^notebooks/",
]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
pythonpath = ["src"]
addopts = [
    "--strict-config",
    "--strict-markers",
    "--disable-warnings",
]
markers = [
    "acceptance: SRD section 16 acceptance suite",
    "integration: end-to-end tests that may use docker compose or mock external services",
    "slow: long-running tests excluded from normal push gates",
    "synthetic: synthetic data allowed only for marked research or unit fixtures",
]
filterwarnings = [
    "error",
]

[tool.coverage.run]
branch = true
source = ["src"]
omit = [
    "src/research/**",
]

[tool.coverage.report]
show_missing = true
skip_covered = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

[tool.importlinter]
root_packages = [
    "app",
    "backtest",
    "data_contract",
    "decision",
    "features",
    "inference",
    "law",
    "research",
    "state",
]

[[tool.importlinter.contracts]]
name = "Domain modules do not import adapters, app, or research"
type = "forbidden"
source_modules = [
    "features",
    "state",
    "law",
    "decision",
    "backtest",
    "inference",
]
forbidden_modules = [
    "app",
    "data_contract",
    "research",
]

[[tool.importlinter.contracts]]
name = "Production modules do not import research"
type = "forbidden"
source_modules = [
    "app",
    "backtest",
    "data_contract",
    "decision",
    "features",
    "inference",
    "law",
    "state",
]
forbidden_modules = [
    "research",
]
